+++ app/views/layout.erb
@@ -17,8 +17,18 @@
       </ul>
     <% end %>
 
+    <% if flash[:notice] %>
+      <div id="notice">
+      <%= flash[:notice] %>
+      </div>
+    <% end %>
+
     <% if current_user %>
       Welcome, <%= current_user.username %>
+      <form method="post" action="/sessions">
+        <input type="hidden" name="_method" value="delete">
+        <input type="submit" value="Sign out">
+      </form>
     <% end %>
 
     <%= yield %>
